flask-api/requirements.txt
Flask==2.2.2
Flask-CORS==3.0.10
Werkzeug==2.3.8
torch==2.2.1
torchvision==0.17.1
numpy==1.26.4
matplotlib==3.8.0
Pillow==10.2.0
python-dotenv==1.0.1
google-generativeai==0.6.0